округление для SVG-ведомости для Схемы

+++ QVertex/tools/coordcatalog.py
@@ -587,10 +587,10 @@ class CatalogData():
                         canvas.text(point[2], insert=((place_v + 6) * mm, (7.5 + place) * mm),
                                     fill='black', font_family='Arial', font_size='9'))
                     canvas.add(
-                        canvas.text(str(math.ceil(point[0])), insert=((30.3 + place_v) * mm, (7.5 + place) * mm), fill='black', font_family='Arial',
+                        canvas.text(str(int(math.ceil(point[0]))), insert=((30.3 + place_v) * mm, (7.5 + place) * mm), fill='black', font_family='Arial',
                                     font_size='9'))
                     canvas.add(
-                        canvas.text(str(math.ceil(point[1])), insert=((55.3 + place_v) * mm, (7.5 + place) * mm), fill='black', font_family='Arial',
+                        canvas.text(str(int(math.ceil(point[1]))), insert=((55.3 + place_v) * mm, (7.5 + place) * mm), fill='black', font_family='Arial',
                                     font_size='9'))
 
                     place += step
